+++ src/parser/inline/extension/label.ts
@@ -8,10 +8,10 @@ export const label: ExtensionParser.LabelParser = line(fmap(build(() =>
   surround(
     '[:',
     match(
-      /^[a-z]+(?:(?:-[0-9]*[a-z][0-9a-z]*)+(?:-0(?:\.0)*)?|-[0-9]+(?:\.[0-9]+)*)/,
+      /^(?:\$|[a-z]+)(?:(?:-[0-9]*[a-z][0-9a-z]*)+(?:-0(?:\.0)*)?|-[0-9]+(?:\.[0-9]+)*)/,
       ([query], rest) =>
         union<ExtensionParser.LabelParser>([link])
-          (`[${query}](#${makeLabel(query)})${rest}`)),
+          (`[\\${query}](#${makeLabel(query)})${rest}`)),
     ']')),
   ([el]) => {
     void el.setAttribute('class', el.getAttribute('href')!.slice(1));